+++ test/mount/lite_test/test_state_2.clj
@@ -2,4 +2,6 @@
   (:require [mount.lite :refer (defstate)]
             [mount.lite-test.test-state-1 :refer (state-1)]))
 
-(defstate state-2 :start (str state-1 " + state-2"))
+(defstate state-2
+  :start (str state-1 " + state-2")
+  :on-cascade :skip)